moved craftmanager to managers
